Change key type of map to uint32.

+++ python/boomer/common/rule_induction.pyx
@@ -170,7 +170,7 @@ cdef class ExactGreedyRuleInduction(RuleInduction):
         cdef uint32[::1] num_conditions_per_comparator = array_uint32(4)
         num_conditions_per_comparator[:] = 0
         # A map that stores the best refinement for each feature
-        cdef unordered_map[intp, Refinement] refinements  # Stack-allocated map
+        cdef unordered_map[uint32, Refinement] refinements  # Stack-allocated map
         # The best refinement of the current rule
         cdef Refinement best_refinement  # Stack-allocated struct
         best_refinement.head = NULL
